move seed data to directory
